requirements.txt
google-cloud-vision==3.7.2
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0
openai==1.23.6
python-dotenv==1.0.1
